Use kebab-case version of `is-approx`
The underscore version of the `is-approx` test function will be removed in
the upcoming initial release of Perl 6, hence the necessecity for this
change.

Note that in one of the comparisons, the `abs_tol` argument needed to be
used, since the comparison was around zero, which is notorious for being
hard to compare with.  Adding the absolute tolerance to the (default)
relative tolerance, allows the comparison to work in this particular case.